Implementação do bucketSort em arquivo + refatoração do bucketSort em lista para funcionar com números inteiros

+++ src/NoLista.java
@@ -1,26 +1,26 @@
 public class NoLista {
-    private double valor;
+    private int valor;
     private NoLista ant, prox;
 
-    public NoLista(double valor, NoLista ant, NoLista prox) {
+    public NoLista(int valor, NoLista ant, NoLista prox) {
         this.valor = valor;
         this.ant = ant;
         this.prox = prox;
     }
 
-    public NoLista(double valor, NoLista ant) {
+    public NoLista(int valor, NoLista ant) {
         this(valor, ant, null);
     }
 
-    public NoLista(double valor) {
+    public NoLista(int valor) {
         this(valor, null, null);
     }
 
-    public double getValor() {
+    public int getValor() {
         return valor;
     }
 
-    public void setValor(double valor) {
+    public void setValor(int valor) {
         this.valor = valor;
     }
